refactor(runner): the allowlist and the job's own types get their own files
runner.go reached 593 lines against the tree's 500-LOC hard cap, so the two
concepts the allowlist work added come out whole rather than being trimmed to
fit.

allowlist.go carries what a catalog entry may call and the two doors that hold
it. job.go carries what a run is GIVEN — goal, grounding, budget, allowlist —
which reads as one idea and was already sitting apart from the loop that consumes
it. runner.go keeps the loop, at 455.

No behaviour change: the split is a move, and the allowlist tests that prove
enforcement at both doors pass unchanged either side of it.
